requirements/common.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-toolbelt==0.0.1
python-dateutil==2.4.2
requests==2.6.0
six==1.9.0
static3==0.5.1
django-pagination
django-tastypie